+++ cohttp-eio/tests/test_server.ml
@@ -5,7 +5,7 @@ let read_body (req, reader) =
   let buf = Buffer.create 0 in
   let fmt = Format.formatter_of_buffer buf in
   Http.Request.pp fmt req;
-  Format.fprintf fmt "\n\n%s%!" (Bytes.unsafe_to_string body);
+  Format.fprintf fmt "\n\n%s%!" body;
   Server.text_response (Buffer.contents buf)
 
 let app (req, reader) =