+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- cohttp-eio: convert to Eio.Buf_read (talex5 #882)
 - cohttp lwt client: Connection cache and explicit pipelining (madroach #853)
 - http: add Http.Request.make and simplify Http.Response.make (bikallem mseri #878)
 - http: add pretty printer functions (bikallem #880)